requirements.txt
click>=8.1.0
openai>=1.0.0
anthropic>=0.18.0
jinja2>=3.1.0
pathlib2>=2.3.0
requests>=2.28.0
rich>=13.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
aiohttp>=3.8.0
asyncio-extras>=1.3.0
fastapi>=0.104.0
uvicorn>=0.24.0
# Local knowledge base dependencies
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
numpy>=1.21.0